Move create_test_data to shared crate
