+++ crates/risc0/host/src/main.rs
@@ -1,11 +1,11 @@
 use leansig_core::{
-    AggregatedVerifier, spec::{self},
+    AggregatedVerifier,
+    spec::{self},
 };
-use leansig_shared::PublicInputs;
+use leansig_shared::{PublicInputs, create_test_data};
 // These constants represent the RISC-V ELF and the image ID generated by risc0-build.
 // The ELF is used for proving and the ID is used for verification.
 use methods::{XMSS_AGGREGATE_ELF, XMSS_AGGREGATE_ID};
-use risc0_host::create_test_data;
 use risc0_zkvm::{ExecutorEnv, default_prover};
 
 fn main() {
@@ -22,7 +22,10 @@ fn main() {
         test_data.public_inputs.spec.clone(),
     );
     assert!(
-        verifier.verify(&test_data.public_inputs.message, &test_data.aggregated_signature),
+        verifier.verify(
+            &test_data.public_inputs.message,
+            &test_data.aggregated_signature
+        ),
         "failed to verify aggregated signature"
     );
