+++ crates/shared/src/lib.rs
@@ -1,4 +1,7 @@
-use leansig_core::{Message, Param, hash::Hash, spec::Spec, AggregatedSignature};
+use leansig_core::{
+    AggregatedSignature, Message, Param, Signer, ValidatorSignature, hash::Hash, spec::Spec,
+};
+use rand::{SeedableRng, rngs::StdRng};
 use serde::{Deserialize, Serialize};
 
 /// Public inputs for RISC0 proof - only this gets committed to the journal
@@ -21,4 +24,72 @@ pub struct PublicInputs {
 pub struct XmssTestData {
     pub public_inputs: PublicInputs,
     pub aggregated_signature: AggregatedSignature,
+}
+
+/// Create test data for XMSS aggregate signatures
+///
+/// # Arguments
+/// * `num_validators` - Number of validators to create
+/// * `spec` - Specification for the signature scheme
+/// * `tree_height` - Height of the XMSS tree (determines number of signatures = 2^height). Default is 13.
+/// * `max_retries` - Maximum number of retries for nonce grinding. Default is 10000.
+/// * `message` - Optional message to sign. Defaults to [42; 32].
+/// * `epoch` - Epoch for signing. Default is 0.
+///
+/// # Returns
+/// An XmssTestData struct containing both public inputs and aggregated signature
+pub fn create_test_data(
+    num_validators: usize,
+    spec: Spec,
+    tree_height: usize,
+    max_retries: usize,
+    message: Option<Message>,
+    epoch: Option<usize>,
+) -> XmssTestData {
+    let message = message.unwrap_or(Message([42; 32]));
+    let epoch = epoch.unwrap_or(0);
+
+    // Calculate lifetime from tree height (2^height)
+    let lifetime = 1 << tree_height;
+
+    let mut validators: Vec<Signer> = (0..num_validators)
+        .map(|i| {
+            Signer::new(
+                StdRng::seed_from_u64(i as u64 + 1),
+                max_retries,
+                spec.clone(),
+                lifetime,
+            )
+        })
+        .collect();
+
+    let validator_roots: Vec<_> = validators.iter().map(|v| v.root.clone()).collect();
+    let validator_params: Vec<_> = validators.iter().map(|v| v.param.clone()).collect();
+
+    // Each validator signs the message
+    let validator_signatures: Vec<ValidatorSignature> = validators
+        .iter_mut()
+        .map(|validator| {
+            let signature = validator.sign(epoch, &message).expect("Failed to sign");
+            ValidatorSignature {
+                epoch,
+                signature,
+                xmss_root: validator.root.clone(),
+                param: validator.param.clone(),
+            }
+        })
+        .collect();
+
+    let aggregated_signature = AggregatedSignature::new(validator_signatures);
+
+    XmssTestData {
+        public_inputs: PublicInputs {
+            message,
+            epoch,
+            validator_roots,
+            validator_params,
+            spec,
+        },
+        aggregated_signature,
+    }
 }
